Collect RxClass Context Data

Get the ancestors of a class specified by source and classId. The ancestors are returned as a graph of nodes and edges.The source, which restricts the results to a particular class type, should be specified when using a MeSH identifier because some MeSH concepts exist in more than one RxClass tree. getClassGraphBySource returns only one tree. To find all trees that contain a concept, use getClassContexts (*_rxclass_context).This resource returns an array of nodes (rxclassMinConceptItem) and an array of edges (rxclassEdge) that link the nodes.

collect_rxclass_context(
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT")
)

Arguments

class_types
Vector of desired classTypes. This vector is also in the order of the API calls will be made. Can be one or more of the following: "MESHPA", "EPC", "MOA", "PE", "PK", "DISEASE", "DISPOS", "CHEM", "SCHEDULE", "STRUCT", "TC", "VA", "ATC1-4".