Extract RxClass Members

Extract RxClass Members

extract_rxclass_members(
  rela_sources = c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
    "SNOMEDCT"),
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT"),
  prior_version = NULL,
  prior_api_version = "3.1.174"
)

Details

RxClass Members represent the relationship between a Class belonging to RxClass Graph and a Drug concept such as an RxNorm Ingredient.